:construction: Separate estimate tab into table and create dialog.
